+++ gearoenix/render/model/rnd-mdl-static.cpp
@@ -11,9 +11,9 @@ gearoenix::render::model::Static::Static(system::File* f, Engine* e, std::shared
     //    LOGE("location: " << f->tell());
     core::Count cc;
     f->read(cc);
-    children.resize(cc);
+    children.resize((size_t) cc);
     for (core::Count i = 0; i < cc; ++i)
-        children[i] = Model::read_child(f, e, c);
+        children[(size_t)i] = Model::read_child(f, e, c);
 }
 
 void gearoenix::render::model::Static::commit(const scene::Scene*)
